AA-76: Add is_allowed property to DatesSummary
This takes over some of the logic in the is_enabled property which
previously checked two things: (1) is the date time-sensitive and
(2) is the date even applicable to this course.

Now, is_enabled is only responsible for time-sensitive checks and
the new is_allowed property checks applicability.

In this way, we can more cleanly separate whether a date block
should show up on the dates sidebar (is_enabled and is_allowed)
and the dates tab (is_allowed).